+++ src/app/i18n/Language.jsx
@@ -67,6 +67,7 @@ export function getLanguage(langCode) {
       'm/s': <u> {translate('m/s')}</u>, // Meters per second
       '°/s': <u> {translate('°/s')}</u>, // Degrees per second
       MW: <u> {translate('MW')}</u>,     // Mega Watts (same as Mega Joules per second)
+      mps: <u>{translate('m/s')}</u>,    // Metres per second
       ps: <u>{translate('/s')}</u>,           // per second
       pm: <u>{translate('/min')}</u>,         // per minute
       s: <u>{translate('secs')}</u>,         // Seconds