+++ ChangeLog.md
@@ -1,3 +1,6 @@
+#2.2.11
+  * Add help system and initial help file
+
 #2.2.10
   * Fix detailed export of module reinforcement packages
   * Use damagedist for exact breakdown of weapons that have more than one type of damage